Do not use public SSH keys coming from the CRs
